src/spotmicroai/requirements.txt
adafruit-circuitpython-motor
adafruit-circuitpython-pca9685
smbus
RPi.GPIO
pick